refactor: extract preview controllers
